authorization done

+++ app/Http/Controllers/IdeaController.php
@@ -13,12 +13,18 @@ class IdeaController extends Controller
         $vilidate = request()->validate([
             "content"=> "required|min:5|max:200",
         ]);
+        $vilidate['user_id'] = auth()->user()->id;
+     
         Idea::create($vilidate);
         return redirect()->route('dashboard')->with("success","Your Success Adding Idea");
     }
 
     public function destroy(Idea $idea)
     {
+        if(auth()->user()->id !== $idea->user_id)
+        {
+            abort(404);
+        }
         $idea->delete();
         return redirect()->route("dashboard")->with("delete-success","Your Success Delete Idea");
     }
@@ -30,12 +36,20 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {   
+        if(auth()->user()->id !== $idea->user_id)
+        {
+            abort(404);
+        }
         $editing = true;
         return view("ideas.show-detail", compact("idea",'editing'));
     }
 
     public function update(Idea $idea)
     {   
+        if(auth()->user()->id !== $idea->user_id)
+        {
+            abort(404);
+        }
         $validate = request()->validate([
             "content"=> "required|min:5|max:200",
         ]);
